docker: Add helper recipe
Using official rust's docker image as base.

Usage is straight forward:

```
    docker run astarte-device-sdk-rust:latest run --example simple -- \
	   --credentials-secret $credentialssecret \
	   --device-id $deviceid \
	   --pairing-url $pairingurl \
	   --realm $realm
```

For the record it can be rebased on ubuntu:21.10
(and cargo should be added to base packages)

My distro RUST setup is not running it out of the box
so I used a more recent one.

For the record the observed error was:

  error[E0599]: no variant or associated item named `OutOfMemory` found for enum `ErrorKind` in the current scope
  -> /root/.cargo/registry/src/github.com-1ecc6299db9ec823/sqlx-core-0.5.10/src/sqlite/connection/establish.rs:107:32

Install example is needed because there is no default installation rule:

  error: there is nothing to install in `astarte_sdk v0.1.0 (/local/src/astarte-device-sdk-rust)`, because it has no binaries

Forwarded: https://github.com/astarte-platform/astarte-device-sdk-rust/pull/31
Relate-to: https://github.com/astarte-platform/astarte-device-sdk-rust/issues/20
Signed-off-by: Philippe Coval <philippe.coval@astrolabe.coop>