Fix handling of int32 payload on LongInteger mapping
In the case of a server-owned longinteger mapping, Astarte sends an
integer number if the value is small enough.
This behavior was not taken into account during the data validation.
A condition was add in the `eq` function that compares the type of the
data received with the mapping type.
Added some tests.

Signed-off-by: Francesco Vaiani <francesco.vaiani@secomind.com>