feat(sdk): add generic for the Store in the SDK and add in memory
Add an in MemoryStore and make the SDK generic over the store. This will
let us keep the values wile the device is running. We do not need to use
the dynamic dispatch anymore and removes the allocation. The default
Store will be the in memory one, there is no need to check if the store
is present.

Signed-off-by: Joshua Chapman <joshua.chapman@secomind.com>