feat(eventloop): separate the event loop from the event channel
When the event loop receives a publish it will spawn a new task to
handle the new payload. This prevents the main loop to block when
handling the new payload. The events are published through a channel
passed when the SDK is initialized. The core SDK data is now held in a
shared struct wrapped in an Arc, this shared struct is then deref from
the device SDK.

Signed-off-by: Joshua Chapman <joshua.chapman@secomind.com>