feat(props): introduce a trait for the property access
Introduce the `PropAccess` trait to access the properties from the
AstarteDeviceSdk. This makes it easier from the outside to use the sdk
with the increasing number of generic arguments, since we can just have
one with the trait for the behaviour we want. Deprecated `get_property`
in favor of the new trait.

Signed-off-by: Joshua Chapman <joshua.chapman@secomind.com>