feat(proto): send custom errors through a gRPC connection to the nodes
The attach procedure now returns a stream of MessageHubEvents, which can
be either a normal event or an error to report to the user/node. Hence,
we define the RecvError enum, including those sdk error to report to the
node that should not cause a panic behavior but should only be reported.

Signed-off-by: Riccardo Gallo <riccardo.gallo@secomind.com>