feat(retention): implement the retention
Implement the retention stored and volatile. The retention Volatile is
stores the publishes in a fixed capacity VecDeque when the connection is
offline. The stored one is implemented by a generic interface in the
store. Packet are stored are marked and removed by the various
connection implementations. Offline packets are resent when starting or
the connection is re-established.

Signed-off-by: Joshua Chapman <joshua.chapman@joshuachp.dev>