feat(sqlite): simplify and improve the SqliteStore
Move from sqlx to the underling dependencies of rusqlite directly. This
will improve how we manage the access to the database and concurrency of
the queries. Separate the connection in write and read queries, the
writer is wrapped in a Mutex, this will prevent deadlock by enforcing a
single writer thread in the application.

Signed-off-by: Joshua Chapman <joshua.chapman@joshuachp.dev>