feat(derive): implement FromEvent for individual
Derive the FromEvent trait for interfaces with aggregation individual.
This will  require the Rust item to be an `enum` with a variant for each
of the mapping's endpoint. To distinguish the aggregation, a name-value
pair was added to the `#[from_event]` attribute. All the derive macros
have also been updated and changed to use the `attributes` instead of
providing a different macro to pars them.

Signed-off-by: Joshua Chapman <joshua.chapman@secomind.com>