fix(rumqttc): use the new version of rumqttc
Fix a concurrency bug in the task that waits for the Acks from astarte
and the one that handles the event loop.

By updating the rumqttc crate other dependencies have been bumped,
mainly rustls. It introduced an error in some tests that generated the
CSR for the Astarte Certificates. Now a self signed CSR/Certificate is
required in the tests (to pass to rustls). This has made it hard to
maintain openssl, since it generates a non commonly used certificate. I
decided to remove it as a dependency, since we are fully using rustls
anyway and it was only for backwards compatibility.

Signed-off-by: Joshua Chapman <joshua.chapman@secomind.com>