fix(sqlite): Set autocheckpoint size to limit the size of the WAL
The WAL size is not limited by the `journal_size_limit` pragma.
That value sets the size at which the journal will be truncated.
To correctly limit the size of a WAL in use we have to set the
`wal_autocheckpoint` pragma to a value lower than the size limit.

Signed-off-by: Luca Arato <luca.arato@secomind.com>